Rename folder to simsz as it is a python module
